move multi-quest buckets and methods to be npc scoped
