+++ zone/client.cpp
@@ -12833,454 +12833,3 @@ void Client::ClientToNpcAggroProcess()
 		LogAggro("Checking Reverse Aggro (client->npc) scanned_npcs ([{}])", npc_scan_count);
 	}
 }
-
-bool Client::CheckHandin(
-	NPC *n,
-	std::map<std::string, uint16> handin,
-	std::map<std::string, uint16> required,
-	std::vector<const EQ::ItemInstance *> items
-)
-{
-	auto h = Handin{};
-	auto r = Handin{};
-
-	std::string log_handin_prefix = fmt::format("[{}] -> [{}]", GetCleanName(), n->GetCleanName());
-
-	// if the npc is a multi-quest npc, we want to re-use our previously set hand-in bucket
-	if (!m_handin_started && m_hand_in.npc && m_hand_in.npc->IsMultiQuestEnabled()) {
-		h = m_hand_in;
-	}
-
-	std::vector<std::pair<const std::map<std::string, uint16>&, Handin&>> datasets = {};
-
-	// if we've already started the hand-in process, we don't want to re-process the hand-in data
-	// we continue to use the originally set hand-in bucket and decrement from it with each successive hand-in
-	if (m_handin_started) {
-		h = m_hand_in;
-	} else {
-		datasets.emplace_back(handin, h);
-	}
-	datasets.emplace_back(required, r);
-
-	const std::string set_hand_in  = "Hand-in";
-	const std::string set_required = "Required";
-	for (const auto &[data_map, current_handin]: datasets) {
-		std::string current_dataset = &current_handin == &h ? set_hand_in : set_required;
-		for (const auto &[key, value]: data_map) {
-			LogNpcHandinDetail("Processing [{}] key [{}] value [{}]", current_dataset, key, value);
-
-			// Handle items
-			if (Strings::IsNumber(key)) {
-				if (const auto *exists = database.GetItem(Strings::ToUnsignedInt(key));
-					exists && current_dataset == set_required) {
-					current_handin.items.emplace_back(HandinEntry{.item_id = key, .count = value});
-				}
-				continue;
-			}
-
-			// Handle money and any other key-value pairs
-			if (key == "platinum") { current_handin.money.platinum = value; }
-			else if (key == "gold") { current_handin.money.gold = value; }
-			else if (key == "silver") { current_handin.money.silver = value; }
-			else if (key == "copper") { current_handin.money.copper = value; }
-		}
-	}
-
-	// pull hand-in items from the item instances
-	if (!m_handin_started) {
-		for (const auto &i: items) {
-			if (!i) {
-				continue;
-			}
-
-			h.items.emplace_back(
-				HandinEntry{
-					.item_id = std::to_string(i->GetItem()->ID),
-					.count = std::max(static_cast<uint16>(i->GetCharges()), static_cast<uint16>(1)),
-					.item = i->Clone()
-				}
-			);
-		}
-	}
-
-	// compare hand-in to required, the item_id can be in any slot
-	bool met_requirement = true;
-	if (h.items.size() == r.items.size() && !h.items.empty() && !r.items.empty()) {
-		for (auto &h_item: h.items) {
-			bool found = false;
-			for (const auto &r_item: r.items) {
-				if (h_item.item_id == r_item.item_id && h_item.count == r_item.count) {
-					found = true;
-					break;
-				}
-			}
-
-			if (!found) {
-				met_requirement = false;
-				break;
-			}
-		}
-	}
-	else {
-		met_requirement = false;
-	}
-
-	// multi-quest
-	if (n && n->IsMultiQuestEnabled()) {
-		for (auto &h_item: h.items) {
-			for (const auto &r_item: r.items) {
-				if (h_item.item_id == r_item.item_id && h_item.count == r_item.count) {
-					h_item.is_multiquest_item = true;
-				}
-			}
-		}
-	}
-
-	for (auto &h_item: h.items) {
-		LogNpcHandinDetail(
-			"{} Hand-in item [{}] ({}) count [{}] is_multiquest_item [{}]",
-			log_handin_prefix,
-			h_item.item->GetItem()->Name,
-			h_item.item_id,
-			h_item.count,
-			h_item.is_multiquest_item
-		);
-	}
-
-	// compare hand-in money to required
-	if (met_requirement) {
-		if (h.money.platinum != r.money.platinum ||
-			h.money.gold != r.money.gold ||
-			h.money.silver != r.money.silver ||
-			h.money.copper != r.money.copper) {
-			met_requirement = false;
-		}
-	}
-
-	// in-case we trigger CheckHand-in multiple times, only set these once
-	if (!m_handin_started) {
-		m_handin_started  = true;
-		m_hand_in         = h;
-		// save original items for logging
-		m_hand_in.original_items = m_hand_in.items;
-		m_hand_in.original_money = m_hand_in.money;
-		m_hand_in.npc = n;
-	}
-
-	// check if npc is guildmaster
-	if (n->IsGuildmaster()) {
-		for (const auto &h_item: m_hand_in.items) {
-			if (!h_item.item) {
-				continue;
-			}
-
-			if (!IsDisciplineTome(h_item.item->GetItem())) {
-				continue;
-			}
-
-			if (n->IsGuildmasterForClient(this)) {
-				TrainDiscipline(h_item.item->GetID());
-				m_hand_in.items.erase(
-					std::remove_if(
-						m_hand_in.items.begin(),
-						m_hand_in.items.end(),
-						[&](const HandinEntry &i) {
-							bool removed = i.item_id == h_item.item_id;
-							if (removed) {
-								LogNpcHandin(
-									"{} Hand-in success, removing discipline tome [{}] from hand-in bucket",
-									log_handin_prefix,
-									i.item_id
-								);
-							}
-							return removed;
-						}
-					),
-					m_hand_in.items.end()
-				);
-			} else {
-				n->Say("You are not a member of my guild. I will not train you!");
-				met_requirement = false;
-				break;
-			}
-		}
-	}
-
-	// print current hand-in bucket
-	LogNpcHandin(
-		"{} > Before processing hand-in | met_requirement [{}] item_count [{}] platinum [{}] gold [{}] silver [{}] copper [{}]",
-		log_handin_prefix,
-		met_requirement,
-		h.items.size(),
-		h.money.platinum,
-		h.money.gold,
-		h.money.silver,
-		h.money.copper
-	);
-
-	LogNpcHandin(
-		"{} >> Handed Items | Item(s) ({}) platinum [{}] gold [{}] silver [{}] copper [{}]",
-		log_handin_prefix,
-		h.items.size(),
-		h.money.platinum,
-		h.money.gold,
-		h.money.silver,
-		h.money.copper
-	);
-
-	int item_count = 1;
-	for (const auto &i: h.items) {
-		LogNpcHandin(
-			"{} >>> item{} [{}] ({}) count [{}]",
-			log_handin_prefix,
-			item_count,
-			i.item->GetItem()->Name,
-			i.item_id,
-			i.count
-		);
-		item_count++;
-	}
-
-
-
-	LogNpcHandin(
-		"{} >> Required Items | Item(s) ({}) platinum [{}] gold [{}] silver [{}] copper [{}]",
-		log_handin_prefix,
-		r.items.size(),
-		r.money.platinum,
-		r.money.gold,
-		r.money.silver,
-		r.money.copper
-	);
-
-	item_count = 1;
-	for (const auto &i: r.items) {
-		auto item = database.GetItem(Strings::ToUnsignedInt(i.item_id));
-
-		LogNpcHandin(
-			"{} >>> item{} [{}] ({}) count [{}]",
-			log_handin_prefix,
-			item_count,
-			item ? item->Name : "Unknown",
-			i.item_id,
-			i.count
-		);
-
-		item_count++;
-	}
-
-	if (met_requirement) {
-		std::vector<std::string> log_entries = {};
-		for (const auto &h_item : h.items) {
-			auto it = std::find_if(r.items.begin(), r.items.end(), [&](const auto &r_item) {
-				return h_item.item_id == r_item.item_id && h_item.count == r_item.count;
-			});
-
-			if (it != r.items.end()) {
-				m_hand_in.items.erase(
-					std::remove_if(
-						m_hand_in.items.begin(),
-						m_hand_in.items.end(),
-						[&](const HandinEntry &i) {
-							bool removed = i.item_id == h_item.item_id && i.count == h_item.count;
-							if (removed) {
-								log_entries.emplace_back(
-									fmt::format(
-										"{} >>> Hand-in success | Removing from hand-in bucket | item [{}] ({}) count [{}]",
-										log_handin_prefix,
-										i.item->GetItem()->Name,
-										i.item_id,
-										i.count
-									)
-								);
-							}
-							return removed;
-						}
-					),
-					m_hand_in.items.end()
-				);
-			}
-		}
-
-		// log successful hand-in items
-		if (!log_entries.empty()) {
-			for (const auto& log : log_entries) {
-				LogNpcHandin("{}", log);
-			}
-		}
-
-		LogNpcHandin(
-			"{} > End of hand-in | met_requirement [{}] item_count [{}] platinum [{}] gold [{}] silver [{}] copper [{}]",
-			log_handin_prefix,
-			met_requirement,
-			m_hand_in.items.size(),
-			m_hand_in.money.platinum,
-			m_hand_in.money.gold,
-			m_hand_in.money.silver,
-			m_hand_in.money.copper
-		);
-		for (const auto &i: m_hand_in.items) {
-			LogNpcHandin(
-				"{} Hand-in success, item [{}] ({}) count [{}]",
-				log_handin_prefix,
-				i.item->GetItem()->Name,
-				i.item_id,
-				i.count
-			);
-		}
-
-		// decrement successful hand-in money from current hand-in bucket
-		if (h.money.platinum > 0 || h.money.gold > 0 || h.money.silver > 0 || h.money.copper > 0) {
-			LogNpcHandin(
-				"{} Hand-in success, removing money p [{}] g [{}] s [{}] c [{}]",
-				log_handin_prefix,
-				h.money.platinum,
-				h.money.gold,
-				h.money.silver,
-				h.money.copper
-			);
-			m_hand_in.money.platinum -= h.money.platinum;
-			m_hand_in.money.gold -= h.money.gold;
-			m_hand_in.money.silver -= h.money.silver;
-			m_hand_in.money.copper -= h.money.copper;
-		}
-	}
-
-	return met_requirement;
-}
-
-void Client::ReturnHandinItems()
-{
-	// player event
-	std::vector<PlayerEvent::HandinEntry> handin_items;
-	PlayerEvent::HandinMoney              handin_money{};
-	std::vector<PlayerEvent::HandinEntry> return_items;
-	PlayerEvent::HandinMoney              return_money{};
-	for (const auto& i : m_hand_in.original_items) {
-		if (i.item && i.item->GetItem()) {
-			handin_items.emplace_back(
-				PlayerEvent::HandinEntry{
-					.item_id = i.item->GetID(),
-					.item_name = i.item->GetItem()->Name,
-					.augment_ids = i.item->GetAugmentIDs(),
-					.augment_names = i.item->GetAugmentNames(),
-					.charges = std::max(static_cast<uint16>(i.item->GetCharges()), static_cast<uint16>(1))
-				}
-			);
-		}
-	}
-
-	// check if any money was handed in
-	if (m_hand_in.original_money.platinum > 0 ||
-		m_hand_in.original_money.gold > 0 ||
-		m_hand_in.original_money.silver > 0 ||
-		m_hand_in.original_money.copper > 0
-		) {
-		handin_money.copper   = m_hand_in.original_money.copper;
-		handin_money.silver   = m_hand_in.original_money.silver;
-		handin_money.gold     = m_hand_in.original_money.gold;
-		handin_money.platinum = m_hand_in.original_money.platinum;
-	}
-
-	bool returned_handin = false;
-	m_hand_in.items.erase(
-		std::remove_if(
-			m_hand_in.items.begin(),
-			m_hand_in.items.end(),
-			[&](auto& i) {
-				if (i.item && i.item->GetItem() && !i.is_multiquest_item) {
-					return_items.emplace_back(
-						PlayerEvent::HandinEntry{
-							.item_id = i.item->GetID(),
-							.item_name = i.item->GetItem()->Name,
-							.augment_ids = i.item->GetAugmentIDs(),
-							.augment_names = i.item->GetAugmentNames(),
-							.charges = std::max(static_cast<uint16>(i.item->GetCharges()), static_cast<uint16>(1))
-						}
-					);
-
-					PushItemOnCursor(*i.item, true);
-					LogNpcHandin("Hand-in failed, returning item [{}]", i.item->GetItem()->Name);
-
-					// Safely delete the item
-					safe_delete(i.item);
-					returned_handin = true;
-					return true; // Mark this item for removal
-				}
-				return false;
-			}
-		),
-		m_hand_in.items.end()
-	);
-
-	// check if any money was handed in
-	if (m_hand_in.money.platinum > 0 ||
-		m_hand_in.money.gold > 0 ||
-		m_hand_in.money.silver > 0 ||
-		m_hand_in.money.copper > 0
-		) {
-		AddMoneyToPP(
-			m_hand_in.money.copper,
-			m_hand_in.money.silver,
-			m_hand_in.money.gold,
-			m_hand_in.money.platinum,
-			true
-		);
-		returned_handin = true;
-		LogNpcHandin(
-			"Hand-in failed, returning money p [{}] g [{}] s [{}] c [{}]",
-			m_hand_in.money.platinum,
-			m_hand_in.money.gold,
-			m_hand_in.money.silver,
-			m_hand_in.money.copper
-		);
-
-		// player event
-		return_money.copper   = m_hand_in.money.copper;
-		return_money.silver   = m_hand_in.money.silver;
-		return_money.gold     = m_hand_in.money.gold;
-		return_money.platinum = m_hand_in.money.platinum;
-	}
-
-	m_has_processed_handin_return = returned_handin;
-
-	if (returned_handin && m_hand_in.npc) {
-		m_hand_in.npc->Say(
-			fmt::format(
-				"I have no need for this {}, you can have it back.",
-				GetCleanName()
-			).c_str()
-		);
-	}
-
-	const bool handed_in_money = (
-		handin_money.platinum > 0 ||
-		handin_money.gold > 0 ||
-		handin_money.silver > 0 ||
-		handin_money.copper > 0
-	);
-	const bool event_has_data_to_record = !handin_items.empty() || handed_in_money;
-
-	if (player_event_logs.IsEventEnabled(PlayerEvent::NPC_HANDIN) && event_has_data_to_record) {
-		auto e = PlayerEvent::HandinEvent{
-			.npc_id = m_hand_in.npc->GetNPCTypeID(),
-			.npc_name = m_hand_in.npc->GetCleanName(),
-			.handin_items = handin_items,
-			.handin_money = handin_money,
-			.return_items = return_items,
-			.return_money = return_money,
-			.is_quest_handin = parse->HasQuestSub(m_hand_in.npc->GetNPCTypeID(), EVENT_TRADE)
-		};
-
-		RecordPlayerEventLogWithClient(this, PlayerEvent::NPC_HANDIN, e);
-	}
-}
-
-void Client::ResetHandin()
-{
-	m_has_processed_handin_return = false;
-	m_handin_started              = false;
-	if (m_hand_in.npc && !m_hand_in.npc->IsMultiQuestEnabled()) {
-		m_hand_in = {};
-	}
-}
