+++ zone/lua_npc.h
@@ -9,6 +9,7 @@ class Lua_Mob;
 class Lua_NPC;
 class Lua_Client;
 struct Lua_NPC_Loot_List;
+class Lua_Inventory;
 
 namespace luabind {
 	struct scope;
@@ -188,6 +189,13 @@ public:
 	void DescribeSpecialAbilities(Lua_Client c);
 	bool IsMultiQuestEnabled();
 	void MultiQuestEnable();
+	bool LuaCheckHandin(
+		Lua_Client c,
+		luabind::adl::object handin_table,
+		luabind::adl::object required_table,
+		luabind::adl::object items_table
+	);
+	void ReturnHandinItems(Lua_Client c);
 };
 
 #endif
