+++ zone/npc.h
@@ -562,11 +562,49 @@ public:
 
 	bool CanPetTakeItem(const EQ::ItemInstance *inst);
 
+	// NPC Hand-in
 	bool IsMultiQuestEnabled() { return m_multiquest_enabled; }
 	void MultiQuestEnable() { m_multiquest_enabled = true; }
 	bool IsGuildmasterForClient(Client *c);
+	bool CheckHandin(
+		Client *c,
+		std::map<std::string, uint16> handin,
+		std::map<std::string, uint16> required,
+		std::vector<const EQ::ItemInstance *> items
+	);
+	void ReturnHandinItems(Client *c);
+	void ResetHandin();
+	bool HasProcessedHandinReturn() { return m_has_processed_handin_return; }
+	bool HandinStarted() { return m_handin_started; }
+
+	struct HandinEntry {
+		std::string            item_id            = "0";
+		uint16                 count              = 0;
+		const EQ::ItemInstance *item              = nullptr;
+		bool                   is_multiquest_item = false;
+	};
+
+	struct HandinMoney {
+		uint32 platinum = 0;
+		uint32 gold     = 0;
+		uint32 silver   = 0;
+		uint32 copper   = 0;
+	};
+
+	struct Handin {
+		std::vector<HandinEntry> original_items = {}; // this is what the player originally handed in, never modified
+		std::vector<HandinEntry> items          = {}; // items can be removed from this set as successful handins are made
+		HandinMoney              original_money = {}; // this is what the player originally handed in, never modified
+		HandinMoney              money          = {}; // money can be removed from this set as successful handins are made
+		NPC                      *npc           = nullptr; // the NPC the player is handing in to
+	};
 
 protected:
+
+	// this is the working handin data from the player
+	// items can be decremented from this as each successful
+	// check is ran in scripts, the remainder is what is returned
+	Handin m_hand_in = {};
 
 	void HandleRoambox();
 
@@ -707,7 +745,10 @@ protected:
 	bool raid_target;
 	bool ignore_despawn; //NPCs with this set to 1 will ignore the despawn value in spawngroup
 
-	bool m_multiquest_enabled = false;
+	// NPC Hand-in
+	bool m_multiquest_enabled          = false;
+	bool m_handin_started              = false;
+	bool m_has_processed_handin_return = false;
 
 private:
 	uint32              m_loottable_id;
